atualizando front

+++ src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
 import { ActivatedRoute, Router, Routes } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { DialogConfirmationComponent } from 'src/app/shared/components/dialog-confirmation/dialog-confirmation.component';
 
 @Component({
   selector: 'app-courses',
@@ -17,6 +18,7 @@ export class CoursesComponent implements OnInit {
   //inicializando a variável aqui..
 
   courses$: Observable<Course[]> | null = null;
+  dialogConfirm: boolean | undefined;
 
   constructor(
     private coursesService: CoursesService,
@@ -47,16 +49,24 @@ export class CoursesComponent implements OnInit {
     this.router.navigate(['edit', course._id], { relativeTo: this.routes })
   }
   onDelete(course: Course) {
-    this.coursesService.delete(course._id).subscribe(
-      () => {
-        this.refresh()
-        this.snackBar.open('Curso removido com sucesso!', 'X', {duration: 3000,
-        verticalPosition: 'top',
-        horizontalPosition: 'center'
-        });
-      },
-        () => this.onError('Erro ao tentar remover curso!')
-    );
+    const dialogRef = this.dialog.open(DialogConfirmationComponent, {
+      data: 'Deseja realmente remover este curso?'
+    });
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.delete(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso!', 'Fechar', {duration: 3000,
+            verticalPosition: 'top',
+            horizontalPosition: 'center'
+            });
+          },
+            () => this.onError('Erro ao tentar remover curso!')
+        );
+      }
+    })
+
   }
 
   refresh() {
@@ -67,6 +77,5 @@ export class CoursesComponent implements OnInit {
         return of([]);
         })
       );
-  }
-
+    }
 }
